% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_rd2markdown.R
\name{add_nl_around_md_blocks}
\alias{add_nl_around_md_blocks}
\title{Add a newline block on either side of block elements}
\usage{
add_nl_around_md_blocks(x)
}
\arguments{
\item{x}{\code{list} of markdown elements, either strings or \code{md_tag} elements}
}
\description{
Add newline elements such that each block element is surrounded by one
newline. Will not add newlines where one already exists.
}
\examples{
x <- list("p1", nl(2), "p2", block("list1"), block("list2"))
add_nl_around_md_blocks(x)
# list("p1", nl(2), "p2", nl(1), block("list1"), nl(1), block("list2"))
#                           ^                      ^

}
