% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_objs.R
\name{block}
\alias{block}
\title{Tag a section as a "block"}
\usage{
block(x = "")
}
\description{
Markdown blocks are sections of markdown which should not be re-wrapped. This
includes things like enumerated lists, itemized lists and tables. Their
individual items might be wrapped, but the block element should always be
surrounded by newlines.
}
\keyword{internal}
