% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_rd2markdown.R
\name{clean_text_newlines}
\alias{clean_text_newlines}
\title{Process whitespace between elements}
\usage{
clean_text_newlines(x)
}
\arguments{
\item{x}{\code{list} of Rd tags}
}
\description{
\code{TEXT} \code{Rd_tag} newlines are discarded, unless they are standalone \code{TEXT}
tags with only a single newline character (and possible non-newline
whitespace) and both preceeded and followed by \code{TEXT} tags.
}
\keyword{internal}
