% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_rd2markdown.R
\name{clean_text_whitespace}
\alias{clean_text_whitespace}
\title{Clean whitespace between text elements}
\usage{
clean_text_whitespace(x)
}
\arguments{
\item{x}{\code{list} of Rd tags}
}
\description{
Within a series of text segments, the first text should have no leading
whitespace, and all but the last element should end with at most one space
character.
}
\examples{
x <- list(" a ", "`b`", " c ", block(), " d   \n", "e   \n", "f")
clean_text_whitespace(x)
# list("a ", "`b`", " c ", block(), "d ", "e ", "f")

x <- list(block(), " a ", "`b`", " c ", block(), " d ", "e ", "f", block())
clean_text_whitespace(x)
# list(block(), "a ", "`b`", " c ", block(), "d ", "e ", "f", block())

}
\keyword{internal}
