% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_utils.R
\name{convert_markdown_headers_to_hash}
\alias{convert_markdown_headers_to_hash}
\title{Convert markdown headers to hash-style}
\usage{
convert_markdown_headers_to_hash(md)
}
\arguments{
\item{md}{A markdown-formatted character value or vector to convert.}
}
\description{
Convert markdown headers to hash-style
}
\keyword{internal}
