% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_to_markdown.R
\name{documentation_to_markdown}
\alias{documentation_to_markdown}
\title{Build a list of documentation files rendered out to markdown}
\usage{
documentation_to_markdown(
  rdfilepaths,
  fragments = c("title", "usage", "description", "details")
)
}
\arguments{
\item{rdfilepaths}{A character vector of Rd documentation file paths.}

\item{fragments}{An optional vector of fragment tag names (such as
"description", "details" or "title") to filter the Rd object tags.}
}
\value{
A named list of character vectors of length one with the markdown content.
The names are dervied from the paths, see the details for more information.
}
\description{
Build a list of documentation files rendered out to markdown
}
\details{
The function provides a vectorized way of extracting multiple documentation
objects at once based on paths to the .Rd files sources. The file path will
be truncated with \code{\link[base]{basename}}, and used as a name for the
output list. Use \code{fragments} parameter whenever you want to limit
extracted documentation only to some parts.
}
\examples{

rd_file_example <- system.file("examples", "rd_file_sample.Rd", package = "rd2markdown")
rd_file_example_2 <- system.file("examples", "rd_file_sample_2.Rd", package = "rd2markdown")

# Extract full documentation for two files
rd2markdown::documentation_to_markdown(c(rd_file_example, rd_file_example_2))

# Limit to particular sections
rd2markdown::documentation_to_markdown(c(rd_file_example, rd_file_example_2),
  fragments = c("details"))

}
\seealso{
\code{\link[rd2markdown]{rd2markdown}} \code{\link[rd2markdown]{get_rd}}
}
