% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_utils.R
\docType{data}
\name{.utils}
\alias{.utils}
\title{Loading unexported helpers from utils}
\format{
An object of class \code{list} of length 3.
}
\usage{
.utils
}
\description{
Loading unexported helpers from utils
}
\keyword{internal}
