% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_rd2markdown.R
\name{map_rd2markdown}
\alias{map_rd2markdown}
\title{Apply rd2markdown across Rd fragments}
\usage{
map_rd2markdown(frags, ..., collapse = NULL)
}
\arguments{
\item{frags}{a list of Rd fragments to apply over}

\item{...}{Additional arguments passed to subsequent \link{rd2markdown} calls.}

\item{collapse}{NULL, the default, does no collapsing. If a character value
is provided, the string results of apply \link{rd2markdown} are collapsed by joining by
the provided value.}
}
\description{
Apply rd2markdown across Rd fragments
}
\seealso{
Other rd2markdown: 
\code{\link{splitRdtag}()}
}
\concept{rd2markdown}
\keyword{internal}
