% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_utils.R
\name{rd_filter_fragments}
\alias{rd_filter_fragments}
\title{Filter an Rd object for fragments by Rd_tag}
\usage{
rd_filter_fragments(x, fragments = c())
}
\arguments{
\item{x}{An Rd object, as produced by \code{tools::parse_Rd}}

\item{fragments}{A \code{character} vector of fragment tags to filter on}
}
\description{
Filter an Rd object for fragments by Rd_tag
}
\keyword{internal}
