% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_rd_sampler.R
\name{rd_sampler}
\alias{rd_sampler}
\title{Rd sampler title}
\arguments{
\item{x}{Rd object}

\item{fragments}{Rd fragments}

\item{...}{Rd sampler ellipsis param}
}
\value{
Rd sampler return
}
\description{
Rd sampler description with \href{https://cran.r-project.org/}{Rd sampler link}, \verb{Rd sampler in-line code}. And Rd dynamic content, \emph{italics text},
\emph{emphasis text}.
}
\details{
Rd sampler details

Rd sampler enumerated list
\enumerate{
\item One
\item Two
\item Three
}

Rd sampler itemized list
\itemize{
\item One
\item Two
\item Three
}\tabular{lll}{
   Rd \tab Sampler \tab Table \cr
   rd \tab sampler \tab table \cr
}


\if{html}{\out{<div class="sourceCode">}}\preformatted{preformatted code
}\if{html}{\out{</div>}}

\eqn{Rd + sampler + inline + equation}

\deqn{Rd * sampler * block * equation}
}
\note{
Rd sampler note
}
\section{Rd sampler subsection}{
Rd sampler subsection text
\subsection{Rd sampler sub-subsection}{

Rd sampler sub-subsection text
}
}

\examples{
print("Hello, World")

}
\references{
R.D. Sampler. (1999)
}
\seealso{
base::print
}
\author{
R.D. Sampler.
}
\concept{rd_samplers}
\keyword{internal}
