% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_utils.R
\name{splitRdtag}
\alias{splitRdtag}
\title{Split a list of Rd fragments at fragments with a specified Rd_tag}
\usage{
splitRdtag(frags, tag)
}
\arguments{
\item{frags}{a list of Rd fragments to split}

\item{tag}{the Rd_tag value upon which the list of Rd fragments should be
split}
}
\description{
Some Rd structures, such as enumerated lists and tables are separated by
placeholder Rd_tags. This function is a convenience function for splitting
those lists into components at the indices of fragments with the provided
tag.
}
\seealso{
Other rd2markdown: 
\code{\link{map_rd2markdown}()}
}
\concept{rd2markdown}
