% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_utils.R
\name{trim_extra_newlines}
\alias{trim_extra_newlines}
\title{Trim non-meaningful markdown newlines}
\usage{
trim_extra_newlines(x)
}
\arguments{
\item{x}{\code{character} vector to process}
}
\description{
We replace one new line sign with two to make sure that proper paragraphs are
always applied. New line without paragraphs is not something used in Rd2
objects and thus we do not account for such situations.
}
\keyword{internal}
