% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_utils.R
\name{trim_extra_newlines}
\alias{trim_extra_newlines}
\title{Trim non-meaningful markdown newlines}
\usage{
trim_extra_newlines(x)
}
\arguments{
\item{x}{\code{character} vector to process}
}
\description{
Markdown does not render single newline character just like it does not
render more than two newline characters. Therefore we strip those unnecessary
newline characters to make them display in a cleaner way.
}
\keyword{internal}
