% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_utils.R
\name{trim_extra_newlines}
\alias{trim_extra_newlines}
\title{Trim non-meaningful markdown newlines}
\usage{
trim_extra_newlines(x)
}
\arguments{
\item{x}{\code{character} vector to process}
}
\description{
Markdown does not render single new line sign just like it does not render
signs multiplied more than two times. Therefore we strim those unnecessary
signs to make them display in a cleaner way.
}
\keyword{internal}
